requirements.txt
datatorch
requests
grpcio
grpcio-tools
numpy
git+https://github.com/jsbroks/dextr-keras.git